backend/requirements.txt
uvicorn
fastapi
python-multipart
pillow
langchain
chromadb
anthropic
openai
python-dotenv
